Fixes after merging with Bartosz

+++ app/src/main/java/com/example/notes/Note.kt
@@ -9,7 +9,7 @@ data class Note(
     @ColumnInfo(name = "text") var text: String,
     @ColumnInfo(name = "brushColor") var brushColor: Int,
     @ColumnInfo(name = "brushWidth") var brushWidth: Float,
-    @ColumnInfo(name = "lowestY") var lowestY: Int,
+    @ColumnInfo(name = "lowestY") var lowestY: Float,
     @ColumnInfo(name = "Paths") var Paths: String,
     @ColumnInfo(name = "id")
     @PrimaryKey(autoGenerate = true) var id: Int? = 0
