+++ app/src/main/java/com/example/notes/NoteDAO.kt
@@ -1,7 +1,6 @@
 package com.example.notes
 
 import android.arch.persistence.room.Dao
-import android.arch.persistence.room.Delete
 import android.arch.persistence.room.Insert
 import android.arch.persistence.room.Query
 
@@ -18,7 +17,7 @@ interface NoteDAO {
     fun insertAll(vararg note : Note)
 
     @Query("update notes set brushColor = :brushColor, brushWidth = :brushWidth, lowestY = :lowestY, Paths = :Paths where id == :idInDB")
-    fun updatePaintNote(idInDB: Int, brushColor: Int, brushWidth: Float, lowestY: Int, Paths: String)
+    fun updatePaintNote(idInDB: Int, brushColor: Int, brushWidth: Float, lowestY: Float, Paths: String)
 
     @Query("update notes set text  = :text where id == :idInDB")
     fun updateTextNote(idInDB: Int, text: String)
